Add `ContributionContext`, separating from `UserContext`
