+++ apps/protocol-frontend/src/components/AttestationsTableShell.tsx
@@ -20,9 +20,10 @@ import MyAttestationsTable from './MyAttestationsTable';
 import ModalWrapper from './ModalWrapper';
 import { GovrnSpinner } from '@govrn/protocol-ui';
 import BulkAttestationModal from './BulkAttestationModal';
+import { useContributions } from '../contexts/ContriubtionContext';
 
 const AttestationsTableShell = () => {
-  const { isDaoContributionLoading, daoContributions } = useUser();
+  const { isDaoContributionLoading, daoContributions } = useContributions();
   const localOverlay = useOverlay();
   const { setModals } = useOverlay();
   const [selectedContributions, setSelectedContributions] = useState<any[]>([]);
