+++ apps/protocol-frontend/src/components/ContributionsTableShell.tsx
@@ -1,20 +1,19 @@
-import { useState, useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import {
-  Container,
-  Stack,
-  Text,
-  Box,
-  Button,
-  Tabs,
-  Tab,
-  TabList,
-  TabPanels,
-  TabPanel,
   Alert,
   AlertDescription,
+  Box,
+  Button,
+  Container,
+  Stack,
+  Tab,
+  TabList,
+  TabPanel,
+  TabPanels,
+  Tabs,
+  Text,
 } from '@chakra-ui/react';
 import { useOverlay } from '../contexts/OverlayContext';
-import { useUser } from '../contexts/UserContext';
 import ModalWrapper from './ModalWrapper';
 import { GovrnSpinner } from '@govrn/protocol-ui';
 import MintModal from './MintModal';
@@ -23,15 +22,15 @@ import PageHeading from './PageHeading';
 import ContributionsTable from './ContributionsTable';
 import ContributionTypesTable from './ContributionTypesTable';
 import EmptyContributions from './EmptyContributions';
+import { useContributions } from '../contexts/ContriubtionContext';
 
 const ContributionsTableShell = () => {
-  const { isUserContributionsLoading, userContributions } = useUser();
+  const { isUserContributionsLoading, userContributions } = useContributions();
   const localOverlay = useOverlay();
   const { setModals } = useOverlay();
   const [selectedContributions, setSelectedContributions] = useState<any>();
   const [selectedContributionsMap, setSelectedContributionsMap] =
     useState<any>();
-  const [selectedDAOs, setSelectedDAOs] = useState<any>();
 
   useEffect(() => {
     if (selectedContributions && selectedContributions.length > 0) {
