+++ apps/protocol-frontend/src/components/DashboardShell.tsx
@@ -8,18 +8,21 @@ import { ControlledSelect } from '@govrn/protocol-ui';
 import { subWeeks } from 'date-fns';
 import ContributionsHeatMap from './ContributionsHeatMap';
 import ContributionsBarChart from './ContributionsBarChart';
+import { useContributions } from '../contexts/ContriubtionContext';
 import { UNASSIGNED } from '../utils/constants';
 
 type UserContributionsDateRangeCountType = {
   count: number;
   date: string;
 };
+
 interface DashboardShellProps {
   user: UIUser | null;
 }
 
 const DashboardShell = ({ user }: DashboardShellProps) => {
-  const { allDaos, userDaos, getUserContributionsCount } = useUser();
+  const { allDaos, userDaos } = useUser();
+  const { getUserContributionsCount } = useContributions();
   const [contributionsCount, setContributionsCount] = useState<
     UserContributionsDateRangeCountType[] | null | undefined
   >([]);
