+++ apps/protocol-frontend/src/components/EditContributionForm.tsx
@@ -13,6 +13,7 @@ import { useUser } from '../contexts/UserContext';
 import { editContributionFormValidation } from '../utils/validations';
 import { UIContribution } from '@govrn/ui-types';
 import { ContributionFormValues } from '../types/forms';
+import { useContributions } from '../contexts/ContriubtionContext';
 
 interface EditContributionFormProps {
   contribution: UIContribution;
@@ -23,7 +24,8 @@ const EditContributionForm = ({
   contribution,
   onClose,
 }: EditContributionFormProps) => {
-  const { updateContribution, userActivityTypes, allDaos } = useUser();
+  const { allDaos, userActivityTypes } = useUser();
+  const { updateContribution } = useContributions();
   const localForm = useForm({
     mode: 'all',
     resolver: yupResolver(editContributionFormValidation),
