+++ apps/protocol-frontend/src/pages/ContributionDetail.tsx
@@ -19,6 +19,7 @@ import ContributionDetailShell from '../components/ContributionDetailShell';
 import NewUserView from '../components/NewUserView';
 import { GOVRN_MOTTO } from '../utils/constants';
 import { UIContribution } from '@govrn/ui-types';
+import { useContributions } from '../contexts/ContriubtionContext';
 
 const UserView = () => {
   return (
@@ -37,7 +38,8 @@ const UserView = () => {
 const ContributionDetails = () => {
   const { isConnected } = useAccount();
   const { id } = useParams();
-  const { userData, getContribution } = useUser();
+  const { userData } = useUser();
+  const { getContribution } = useContributions();
   const { isAuthenticated } = useAuth();
   const [contribution, setContribution] = useState<UIContribution>(
     {} as UIContribution,
